Move `send_generic_email` function to the email utility file
